Added secure WebSocket and external config
Gave the ability to enable encrypted Websocket so Firefox Users with HTTPS connection to tumblr are able to connect as well. Encrypted Websocket is enabled by default and you can either provide a Key/Certificate Pair or a PFX-File. You have to create them yourself.
Also, the config has now been moved into an external file which is created on startup if it doesn't exist and loaded every time the server boots up.